+++ internal/provider/data_source_iosxe_interface_ethernet_test.go
@@ -230,7 +230,7 @@ func testAccDataSourceIosxeInterfaceEthernetConfig() string {
 		config += `	ip_dhcp_snooping_trust = true` + "\n"
 	}
 	config += `	negotiation_auto = false` + "\n"
-	config += `	service_instance = [{` + "\n"
+	config += `	service_instances = [{` + "\n"
 	config += `	}]` + "\n"
 	config += `	ip_flow_monitors = [{` + "\n"
 	config += `		name = "MON1"` + "\n"