+++ docs/guides/changelog.md
@@ -11,7 +11,7 @@ description: |-
 
 - Add `mld_snooping` and `mld_snooping_querier` attributes to the `iosxe_system` resource and data sources
 - Add `iosxe_interface_bdi` resource and data sources
-- Add `service_instance` attribute to `iosxe_interface_ethernet` resource and data sources
+- Add `service_instances` attribute to `iosxe_interface_ethernet` resource and data sources
 - Add `iosxe_bridge_domain` resource and data sources
 - Add `trunk_allowed_vlans_all`, `trunk_allowed_vlans_none`, `trunk_allowed_vlans_add`, `trunk_allowed_vlans_except`, and `trunk_allowed_vlans_remove` attributes to `iosxe_interface_switchport` resource and data source using non-deprecated YANG vlan-v2 paths
 - Add `trunk_allowed_vlans_legacy` and `trunk_allowed_vlans_none_legacy` attributes to `iosxe_interface_switchport` resource and data source for backward compatibility with deprecated YANG paths